+++ lib/SimpleCacheAPCu.php
@@ -2,28 +2,18 @@
 declare(strict_types = 1);
 
 /**
- * An implementation of the PSR-16 SimpleCache Interface for APCu
+ * An implementation of the PSR-16 SimpleCache Interface for APCu.
  *
- * @package AWonderPHP\SimpleCacheAPCu
+ * @package AWonderPHP/SimpleCacheAPCu
  * @author  Alice Wonder <paypal@domblogger.net>
  * @license https://opensource.org/licenses/MIT MIT
  * @link    https://github.com/AliceWonderMiscreations/SimpleCacheAPCu
  */
-/*
- +-------------------------------------------------------+
- |                                                       |
- | Copyright (c) 2018 Alice Wonder Miscreations          |
- |  May be used under terms of MIT license               |
- |                                                       |
- +-------------------------------------------------------+
- | Purpose: PSR-16 APCu Interface                        |
- +-------------------------------------------------------+
-*/
 
 namespace AWonderPHP\SimpleCacheAPCu;
 
 /**
- * An implementation of the PSR-16 SimpleCache Interface for APCu
+ * An implementation of the PSR-16 SimpleCache Interface for APCu.
  *
  * This class implements the [PHP-FIG PSR-16](https://www.php-fig.org/psr/psr-16/)
  *  interface for a cache class.
@@ -31,272 +21,15 @@ namespace AWonderPHP\SimpleCacheAPCu;
  * It needs PHP 7.1 or newer and obviously the [APCu PECL](https://pecl.php.net/package/APCu) extension.
  *  I am not sure of the minimum APCu version, I am using 5.1.9 myself at the moment.
  */
-class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
+class SimpleCacheAPCu extends \AWonderPHP\SimpleCache\SimpleCache implements \Psr\SimpleCache\CacheInterface
 {
     /**
-     * When false, the class does not attempt to write/read from cache
+     * A wrapper for the actual fetch from the cache.
      *
-     * @var bool
-     */
-    protected $enabled = false;
-
-    /**
-     * When false, the class is somewhat tolerant of incorrect parameter type
-     * and will recast if safe to do so.
+     * @param string $realKey The internal key used with APCu.
+     * @param mixed  $default The value to return if a cache miss.
      *
-     * @var bool
-     */
-    protected $strictType = false;
-
-    /**
-     * The salt to use when generating internal key used with APCu
-     *
-     * @var string
-     */
-    protected $salt = '6Dxypt3ePw2SM2zYzEVAFkDBQpxbk16z1';
-
-    /**
-     * The prefix to use with the internal key used with APCu
-     *
-     * @var string
-     */
-    protected $webappPrefix = 'DEFAULT_';
-
-    /**
-     * The default TTL in seconds. 0 means as long as possible to APCu
-     *
-     * @var int
-     */
-    protected $defaultSeconds = 0;
-
-    /**
-     * Checks whether or not a parameter is of the iterable pseudo-type
-     *
-     * @param mixed $arg The parameter to be checked
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException::typeNotIterable($arg)
-     *
-     * @return void
-     */
-    protected function checkIterable($arg): void
-    {
-        if (! is_iterable($arg)) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::typeNotIterable($arg);
-        }
-    }
-
-    /**
-     * Creates hash substring to use in internal cache key.
-     *
-     * This class obfuscates the user supplied cache keys by using a substring
-     * of the hex representation of a hash of that key. This function creates
-     * the hex representation of the hash and grabs a 16 character substring.
-     *
-     * @param string $key The user defined key to hash
-     *
-     * @return string
-     */
-    protected function weakHash($key): string
-    {
-        $key = $this->salt . $key;
-        $key = hash('ripemd160', $key);
-        // 16^16 should be enough of the hash to avoid collisions
-        return substr($key, 17, 16);
-    }
-
-    /**
-     * Takes user supplied key and creates internal cache key.
-     *
-     * This function takes the user defined cache key, gets the substring of a
-     * hash from the weakHash function, and appends that substring to the WebApp
-     * prefix string to create the actual key that will be used with APCu.
-     *
-     * @param string $key The user defined cache key to hash
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress RedundantConditionGivenDocblockType
-     *
-     * @return string
-     */
-    protected function adjustKey($key): string
-    {
-        if (! $this->strictType) {
-            $invalidTypes = array('array', 'object', 'boolean', 'NULL');
-            $type = gettype($key);
-            if (! in_array($type, $invalidTypes)) {
-                $key = (string)$key;
-            }
-        }
-        if (! is_string($key)) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::keyTypeError($key);
-        }
-        $key = trim($key);
-        if (strlen($key) === 0) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::emptyKey();
-        }
-        if (strlen($key) > 255) {
-            // key should not be larger
-            //  than 255 character
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::keyTooLong($key);
-        }
-        if (preg_match('/[\[\]\{\}\(\)\/\@\:\\\]/', $key) !== 0) {
-            // PSR-16 says those characters not allowed
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::invalidKeyCharacter($key);
-        }
-        $key = $this->webappPrefix . $this->weakHash($key);
-        return $key;
-    }
-
-    /**
-     * Sets the prefix (namespace) for the internal keys
-     *
-     * @param string $str The string to use as internal key prefix.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress RedundantConditionGivenDocblockType
-     *
-     * @return void
-     */
-    protected function setWebAppPrefix($str): void
-    {
-        $type = gettype($str);
-        if (! is_string($str)) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::cstrTypeError($str, 'WebApp Prefix');
-        }
-        $str = strtoupper(trim($str));
-        if (strlen($str) < 3) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::webappPrefixTooShort($str);
-        }
-        if (strlen($str) > 32) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::webappPrefixTooLong($str);
-        }
-        if (preg_match('/[^A-Z0-9]/', $str) !== 0) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::webappPrefixNotAlphaNumeric($str);
-        }
-        $this->webappPrefix = $str . '_';
-    }
-
-    /**
-     * Sets the salt to use when generating the internal keys
-     *
-     * @param string $str The string to use as the salt when creating the internal key prefix.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress RedundantConditionGivenDocblockType
-     *
-     * @return void
-     */
-    protected function setHashSalt($str): void
-    {
-        $type = gettype($str);
-        if (! is_string($str)) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::cstrTypeError($str, 'Salt');
-        }
-        $str = trim($str);
-        if (strlen($str) < 8) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::saltTooShort($str);
-        }
-        $this->salt = $str;
-    }
-
-    /**
-     * Converts a \DateInterval object to seconds
-     *
-     * @param \DateInterval $interval The date interval to be converted into seconds
-     *
-     * @return int The number of seconds corresponding to the DateInterval
-     */
-    protected function dateIntervalToSeconds($interval): int
-    {
-        $now = time();
-        $dt = new \DateTime();
-        $dt->add($interval);
-        $ts = $dt->getTimestamp();
-        $diff = $ts - $now;
-        return $diff;
-    }
-
-    /**
-     * Generates Time To Live parameter to use with APCu.
-     *
-     * This function takes either NULL, an integer or a string. When supplied
-     * with an integer, if it is less than the current seconds from UNIX Epoch
-     * it is treated as desired seconds the record should last. If it is larger
-     * it assumed it is an expiration time and then calculates the corresponding
-     * TTL. When fed a string, the `strtotime()` function is used to turn the
-     * string into a UNIX seconds from Epoch expiration, and it then calculates
-     * the corresponding TTL. When fed NULL, it uses the class default TTL.
-     *
-     * @param null|int|string|\DateInterval $ttl The length to cache or the expected expiration.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress RedundantConditionGivenDocblockType
-     * @psalm-suppress RedundantCondition
-     *
-     * @return int
-     */
-    protected function ttlToSeconds($ttl): int
-    {
-        if (is_null($ttl)) {
-            return $this->defaultSeconds;
-        }
-        if (is_object($ttl)) {
-            if ($ttl instanceof \DateInterval) {
-                $seconds = $this->dateIntervalToSeconds($ttl);
-                if ($seconds < 0) {
-                    throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::dateIntervalInPast();
-                }
-                return $seconds;
-            } else {
-                throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::ttlTypeError($ttl);
-            }
-        }
-        if (! $this->strictType) {
-            if (is_numeric($ttl)) {
-                $ttl = intval($ttl, 10);
-            }
-        }
-        $type = gettype($ttl);
-        if (! in_array($type, array('integer', 'string'))) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::ttlTypeError($ttl);
-        }
-        $now = time();
-        if (is_int($ttl)) {
-            $seconds = $ttl;
-            if ($seconds > $now) {
-                return ($seconds - $now);
-            }
-            if ($seconds < 0) {
-                throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::negativeTTL($seconds);
-            }
-            return $seconds;
-        }
-        // hope it is a date string
-        if ($seconds = strtotime($ttl, $now)) {
-            if ($seconds > $now) {
-                return ($seconds - $now);
-            } else {
-                throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::dateStringInPast($ttl);
-            }
-        }
-        throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::invalidTTL($ttl);
-    }
-
-    /**
-     * A wrapper for the actual fetch from the cache
-     *
-     * @param string $realKey The internal key used with APCu
-     * @param mixed  $default The value to return if a cache miss
-     *
-     * @return mixed The value in the cached key => value pair, or $default if a cache miss
+     * @return mixed The value in the cached key => value pair, or $default if a cache miss.
      */
     protected function cacheFetch($realKey, $default)
     {
@@ -305,136 +38,41 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
             return $return;
         }
         return $default;
-    }
+    }//end cacheFetch()
+
 
     /**
-     * A wrapper for the actual store of key => value pair in the cache
+     * A wrapper for the actual store of key => value pair in the cache.
      *
-     * @param string                        $realKey The internal key used with APCu
-     * @param mixed                         $value   The value to be stored
+     * @param string                        $realKey The internal key used with APCu.
+     * @param mixed                         $value   The value to be stored.
      * @param null|int|string|\DateInterval $ttl     The TTL value of this item.
      *
-     * @return bool Returns True on success, False on failure
+     * @return bool Returns True on success, False on failure.
      */
     protected function cacheStore($realKey, $value, $ttl): bool
     {
         $seconds = $this->ttlToSeconds($ttl);
         return apcu_store($realKey, $value, $seconds);
-    }
+    }//end cacheStore()
+
 
     /**
-     * A wrapper for the actual delete of a key => value pair in the cache
+     * A wrapper for the actual delete of a key => value pair in the cache.
      *
-     * @param string $realKey The key for the key => value pair to be removed from the cache
+     * @param string $realKey The key for the key => value pair to be removed from the cache.
      *
-     * @return bool Returns True on success, False on failure
+     * @return bool Returns True on success, False on failure.
      */
     protected function cacheDelete($realKey): bool
     {
         return apcu_delete($realKey);
-    }
+    }//end cacheDelete()
 
-    /**
-     * Sets the default cache TTL in seconds.
-     *
-     * @param int|\DateInterval $ttl The default TTL to cache entries
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @psalm-suppress RedundantConditionGivenDocblockType
-     * @psalm-suppress RedundantCondition
-     *
-     * @return void
-     */
-    public function setDefaultSeconds($ttl): void
-    {
-        $type = gettype($ttl);
-        switch ($type) {
-            case "integer":
-                $seconds = $ttl;
-                break;
-            case "object":
-                if ($ttl instanceof \DateInterval) {
-                    $seconds = $this->dateIntervalToSeconds($ttl);
-                } else {
-                    throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::defaultTTL($ttl);
-                }
-                break;
-            default:
-                if (! $this->strictType) {
-                    if (is_numeric($ttl)) {
-                        $seconds = intval($ttl);
-                    }
-                }
-        }
-        if (! isset($seconds)) {
-            throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::defaultTTL($ttl);
-        }
-        if ($seconds < 0) {
-            throw \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException::negativeDefaultTTL($seconds);
-        }
-        $this->defaultSeconds = $seconds;
-    }
-
-    /**
-     * Fetches a value from the cache.
-     *
-     * @param string $key     The unique key of this item in the cache.
-     * @param mixed  $default (optional) Default value to return if the key does not exist.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @return mixed The value of the item from the cache, or $default in case of cache miss.
-     */
-    public function get($key, $default = null)
-    {
-        $realKey = $this->adjustKey($key);
-        if ($this->enabled) {
-            return $this->cacheFetch($realKey, $default);
-        }
-        return $default;
-    }
-
-    /**
-     * Persists data in the cache, uniquely referenced by a key with an optional expiration TTL time.
-     *
-     * @param string                        $key   The key of the item to store.
-     * @param mixed                         $value The value of the item to store, must be
-     *                                             serializable.
-     * @param null|int|string|\DateInterval $ttl   (optional) The TTL value of this item.
-     *
-     * @return bool True on success and false on failure.
-     */
-    public function set($key, $value, $ttl = null): bool
-    {
-        $realKey = $this->adjustKey($key);
-        if ($this->enabled) {
-            return $this->cacheStore($realKey, $value, $ttl);
-        }
-        return false;
-    }
-
-    /**
-     * Delete an item from the cache by its unique key.
-     *
-     * @param string $key The unique cache key of the item to delete.
-     *
-     * @return bool True if the item was successfully removed. False if there was an error.
-     */
-    public function delete($key): bool
-    {
-        if ($this->enabled) {
-            $realKey = $this->adjustKey($key);
-            return $this->cacheDelete($realKey);
-        }
-        return false;
-    }
 
     /**
      * Wipes clean the entire cache's keys. This implementation only wipes for matching
-     * webappPrefix (custom NON PSR-16 feature set during constructor)
+     * webappPrefix (custom NON PSR-16 feature set during constructor).
      *
      * @return bool True on success and false on failure.
      */
@@ -465,7 +103,8 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
             }
         }
         return false;
-    }
+    }//end clear()
+
 
     /**
      * Wipes clean the entire cache's keys regardless of webappPrefix.
@@ -482,100 +121,8 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
             }
         }
         return $return;
-    }
+    }//end clearAll()
 
-    /**
-     * Obtains multiple cache items by their unique keys.
-     *
-     * @param iterable $keys    A list of keys that can obtained in a single operation.
-     * @param mixed    $default Default value to return for keys that do not exist.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @return array A list of key => value pairs. Cache keys that do not exist or are
-     *                                                stale will have $default as value.
-     */
-    public function getMultiple($keys, $default = null): array
-    {
-        $this->checkIterable($keys);
-        $return = array();
-        foreach ($keys as $userKey) {
-            if (! is_string($userKey)) {
-                throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::iterableKeyMustBeString($userKey);
-            }
-            $value = $default;
-            if ($this->enabled) {
-                $realKey = $this->adjustKey($userKey);
-                $value = $this->cacheFetch($realKey, $default);
-            }
-            $return[$userKey] = $value;
-        }
-        return $return;
-    }
-
-    /**
-     * Persists a set of key => value pairs in the cache, with an optional TTL.
-     *
-     * @param iterable                      $pairs A list of key => value pairs for multiple
-     *                                             set operation.
-     * @param null|int|string|\DateInterval $ttl   (optional) The TTL value of this item.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @return bool True on success and false on failure.
-     */
-    public function setMultiple($pairs, $ttl = null): bool
-    {
-        if (! $this->enabled) {
-            return false;
-        }
-        $this->checkIterable($pairs);
-        $arr = array();
-        foreach ($pairs as $key => $value) {
-            if (! is_string($key)) {
-                throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::iterableKeyMustBeString($key);
-            }
-            $realKey = $this->adjustKey($key);
-            $arr[$realKey] = $value;
-        }
-        foreach ($arr as $realKey => $value) {
-            if (! $this->cacheStore($realKey, $value, $ttl)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Deletes multiple cache items in a single operation.
-     *
-     * @param iterable $keys A list of string-based keys to be deleted.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @return bool True if the items were successfully removed. False if there was an error.
-     */
-    public function deleteMultiple($keys): bool
-    {
-        if (! $this->enabled) {
-            return false;
-        }
-        $return = true;
-        $this->checkIterable($keys);
-        foreach ($keys as $userKey) {
-            if (! is_string($userKey)) {
-                throw \AWonderPHP\SimpleCacheAPCu\StrictTypeException::iterableKeyMustBeString($userKey);
-            }
-            $realKey = $this->adjustKey($userKey);
-            if (! $this->cacheDelete($realKey)) {
-                $return = false;
-            }
-        }
-        return $return;
-    }
 
     /**
      * Determines whether an item is present in the cache.
@@ -596,22 +143,8 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
             return apcu_exists($key);
         }
         return false;
-    }
+    }//end has()
 
-    /**
-     * Returns the actual internal key being used with APCu. Needed for unit testing.
-     *
-     * @param string $key The key you wanted translated to the APCu key.
-     *
-     * @throws \AWonderPHP\SimpleCacheAPCu\StrictTypeException
-     * @throws \AWonderPHP\SimpleCacheAPCu\InvalidArgumentException
-     *
-     * @return string
-     */
-    public function getRealKey($key): string
-    {
-        return $this->adjustKey($key);
-    }
 
     /**
      * Class constructor function. Takes three arguments with defaults.
@@ -653,8 +186,7 @@ class SimpleCacheAPCu implements \Psr\SimpleCache\CacheInterface
             $this->enabled = true;
         }
         $this->strictType = $strictType;
-    }
-}
+    }//end __construct()
+}//end class
 
-// Dear PSR-2: You can take my closing PHP tag when you can pry it from my cold dead fingers.
 ?>